// app/blackjack/types.ts
export type Suit = "hearts" | "diamonds" | "clubs" | "spades";
export type Rank =
  | "A"
  | "2"
  | "3"
  | "4"
  | "5"
  | "6"
  | "7"
  | "8"
  | "9"
  | "10"
  | "J"
  | "Q"
  | "K";

export interface Card {
  suit: Suit;
  rank: Rank;
  faceUp: boolean;
}

export type GameStatus =
  | "betting"
  | "playing"
  | "dealer-turn"
  | "player-bust"
  | "dealer-bust"
  | "player-blackjack"
  | "player-win"
  | "dealer-win"
  | "push";